Actor velocity to float

+++ src/Engine/Objects/Actor.h
@@ -211,7 +211,7 @@ class Actor {
     uint16_t height = 128;
     uint16_t moveSpeed = 200;
     Vec3f pos;
-    Vec3i velocity;
+    Vec3f velocity;
     uint16_t yawAngle = 0;
     uint16_t pitchAngle = 0;
     int sectorId = 0;
